requirements.txt
streamlit>=1.24.0
tensorflow>=2.10.0
torch>=1.13.0
opencv-python>=4.8.0
Pillow>=9.0.0
numpy>=1.23.0
pyrebase4>=4.5.0
seaborn>=0.11.2
